ImageSelector now uses the SpriteKey class and can be used to select individual sprites in an atlas. MissileRigEditorWidget now displays the selected sprite

+++ SFMLSpaceGame/UI/WidgetHelpers.cpp
@@ -17,7 +17,7 @@ namespace WidgetHelpers
 	{
 		auto mousePos = sf::Mouse::getPosition(*GAME_WINDOW);
 		auto canvasPos = widget->GetAbsolutePosition();
-		return sf::Vector2f(mousePos.x, mousePos.y) - canvasPos;
+		return sf::Vector2f(static_cast<float>(mousePos.x), static_cast<float>(mousePos.y)) - canvasPos;
 	}
 
 	bool MouseInWidget(std::shared_ptr<sfg::Widget> widget)
